fix(Post): Fix directory structure
