+++ src/components/Post/PostDetail.tsx
@@ -4,8 +4,8 @@ import Link from "next/link";
 import React, { useEffect, useRef } from "react";
 import ReactDOM from "react-dom";
 import { useIsomorphicLayoutEffect } from "../../lib/hooks/useIsomarphicLayoutEffect";
-import { PostFragment } from "../Post/__generated__/index.generated";
-import { useGetCommentsQuery } from "./__generated__/index.generated";
+import { PostFragment } from "./__generated__/Post.generated";
+import { useGetCommentsQuery } from "./__generated__/PostDetail.generated";
 
 const AVATAR_URL = "https://dummyimage.com/88x88/b3b3b3/ffffff";
 const AVATAR_URL_36 = "https://dummyimage.com/36x36/b3b3b3/ffffff";
